A visual pattern to describe that ideias mayB4real
If you feel life making sound experiments to see the visualization of mathematical concepts to have some fun and relax. You will need:
Ableton Live Suite Liscence;
Izotope (Ozone8) TIP << if on a budget Splice has a deal to pay monthly instead of in full for Izotope Ozone 8 liscence;

that is about what will be needed to complete the task.